1)add command line argument control; 2)using a common header
