+++ aws/cl_zcash/software/runtime/makefile_ecdsa
@@ -22,19 +22,20 @@ INCLUDES += -I ./include
 CC = g++
 CFLAGS = -DCONFIG_LOGLEVEL=4 -g -Wall $(INCLUDES) -lstdc++ -std=c++11
 
-LDLIBS = -lfpga_mgmt -lrt -lpthread
+LDLIBS = -lfpga_mgmt -lrt -lpthread -lcrypto -lssl
 
 SRC = zcash_fpga.cpp ecdsa_test.cpp ${SDK_DIR}/userspace/utils/sh_dpi_tasks.c
+
 OBJ = $(SRC:.c=.o)
 BIN = ecdsa_test
 
 all: $(BIN) check_env
 
 $(BIN): $(OBJ)
-        $(CC) $(CFLAGS) -o $@ $^ $(LDFLAGS) $(LDLIBS)
+	$(CC) $(CFLAGS) -o $@ $^ $(LDFLAGS) $(LDLIBS)
 
 clean:
-        rm -f *.o $(BIN)
+	rm -f *.o $(BIN)
 
 check_env:
 ifndef SDK_DIR
